Adding Svn tags and branches.
Svn does not formally support tags and branches. Instead there
is a convention that tags live in the tags/ directory and
branches live in the branches/ directory. The latest work is
in the trunk/ directory. Since these are all conventions rather
than part of the VCS we assume the url being checked out is to
the right place. Tags and Braches return empty arrays. We
assume the implementor handles the conventions properly for their
situation.