Fix svn tests
Note: the test to detect a change in underlying VCS was removed.
This test detected when SVN was retrieved from GitHub but it
supplied a .git directory which was locally detected as Git instead
of svn. Since GitHub has dropped svn support there is no place to
test this with integration tests.

Signed-off-by: Matt Farina <matt.farina@suse.com>